move css to separate file
