+++ Card.js
@@ -1,62 +1,7 @@
 import React from 'react';
-import {StyleSheet, Text, View, Image, Dimensions, Button} from 'react-native';
+import {Text, View, Image, Dimensions, Button} from 'react-native';
 
-const width = Dimensions.get('window').width; //full width
-const height = Dimensions.get('window').height; //full height
-
-
-const styles = StyleSheet.create({
-  hero: {
-    paddingTop: 50,
-    backgroundColor: '#66b34e',
-    flexDirection: 'row',
-    flex: 3,
-  },
-  lookingFor: {
-    backgroundColor: '#e3da74',
-    flex: 1,
-  },
-  profileData: {
-    backgroundColor: '#66b34e',
-    flex: 3,
-  },
-  actions: {
-    backgroundColor: '#e3da74',
-    alignSelf: 'stretch',
-    justifyContent: 'center',
-    padding: 50,
-    flexDirection: 'row',
-    flex: 1,
-  },
-  card: {
-    flex: 1,
-    width: width,
-    alignSelf: 'stretch',
-    alignItems: 'center',
-    borderRadius: 5,
-    overflow: 'hidden',
-    borderColor: 'grey',
-    backgroundColor: 'white',
-    borderWidth: 1,
-    elevation: 1,
-  },
-  thumbnail: {
-    alignSelf: 'center',
-    width: 200,
-    height: 200,
-    borderRadius: 100,
-  },
-  text: {
-    fontSize: 20,
-    paddingTop: 10,
-    paddingBottom: 5,
-  },
-  noMoreCards: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  }
-})
+import styles from './css/card.js'
 
 class Card extends React.Component {
   handleUndo() {
@@ -69,6 +14,7 @@ class Card extends React.Component {
 
     return (
       <View style={styles.card}>
+
         <View style={styles.hero}>
           <View style={{flex: 1, backgroundColor: 'powderblue'}} >
           </View>
@@ -78,6 +24,7 @@ class Card extends React.Component {
           <View style={{flex: 1, backgroundColor: 'powderblue'}} >
           </View>
         </View>
+
         <View style={styles.lookingFor}>
           <Text style={styles.text} ellipsizeMode={'tail'} numberOfLines={3} >{this.props.mentorLookingFor}</Text>
         </View>
